[compiler] Be explicit in what constructor to use.
When using the Storage constructor, there's a type conflict between
uint64_t and uintptr_t. We just need to be explicit on which one to use.

+++ src/compiler/turboshaft/assembler.h
@@ -2522,7 +2522,8 @@ class AssemblerOpInterface : public Next {
 
   V<Word32> RelocatableWasmIndirectCallTarget(uint32_t function_index) {
     return ReduceIfReachableConstant(
-        ConstantOp::Kind::kRelocatableWasmIndirectCallTarget, function_index);
+        ConstantOp::Kind::kRelocatableWasmIndirectCallTarget,
+        static_cast<uint64_t>(function_index));
   }
 
   V<Context> NoContextConstant() {
